refactor(workers): split lifecycle options
