+++ src/gateway/worker-environments/service-validation.ts
@@ -5,8 +5,10 @@ import { WorkerMachineOptionsSchema } from "../../../packages/gateway-protocol/s
 import {
   WorkerProviderError,
   type WorkerDesktopEndpoint,
+  type WorkerExecutionMode,
   type WorkerLease,
   type WorkerLeaseStatus,
+  type WorkerProvider,
   type WorkerMachineOption,
   type WorkerSshEndpoint,
 } from "../../plugins/types.js";
@@ -82,6 +84,20 @@ export function requireWorkerLeaseStatus(value: unknown): WorkerLeaseStatus {
     throw new Error("Worker provider returned an invalid inspection result");
   }
   return { status };
+}
+
+export function resolveWorkerLeaseModeError(
+  provider: WorkerProvider,
+  lease: WorkerLease,
+): WorkerProviderError | undefined {
+  const modes = provider.supportedExecutionModes;
+  const executionMode: WorkerExecutionMode | undefined = modes?.length === 1 ? modes[0] : undefined;
+  const leaseMode: WorkerExecutionMode = lease.node ? "worker-turn" : "remote-exec";
+  return !executionMode || executionMode === leaseMode
+    ? undefined
+    : new WorkerProviderError(
+        `${executionMode} providers must return a ${executionMode === "worker-turn" ? "node" : "SSH"} lease`,
+      );
 }
 
 export function requireWorkerLease(value: unknown): WorkerLease {
